pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cube-alchemy"
version = "0.1.13"
description = "A lightweight, queryable hypercube engine for multidimensional data analysis"
readme = "READMEPIP.md"
authors = [
    {name = "Juan C. Del Monte", email = "delmontejuan92@gmail.com"},
]
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "networkx>=3.5,<4.0",
    "matplotlib>=3.10.3,<4.0",
    "pandas>=2.0.0,<3.0",
    "pyyaml>=6.0.0,<7.0",
]

[project.urls]
"Docs" = "https://cube-alchemy.info"
"Repository" = "https://github.com/cube-alchemy/cube-alchemy"
"Bug Tracker" = "https://github.com/cube-alchemy/cube-alchemy/issues"

[tool.setuptools]
packages = [
    "cube_alchemy",
    "cube_alchemy.core",
    "cube_alchemy.core.hypercube_building_classes",
    "cube_alchemy.model_definitions",
    "cube_alchemy.model_definitions.sources",
    "cube_alchemy.model_definitions.repository",
    "cube_alchemy.model_definitions.validation",
]